+++ genbase/__init__.py
@@ -7,6 +7,7 @@ import numpy as np
 import srsly
 
 from genbase.data import import_data, train_test_split
+from genbase.decorator import add_calldict
 from genbase.internationalization import (LOCALE_MAP, get_locale, set_locale,
                                           translate_list, translate_string)
 from genbase.mixin import CaseMixin, SeedMixin
@@ -139,7 +140,7 @@ class MetaInfo(Configurable):
 
     def to_config(self):
         content = self.content if hasattr(self, 'content') \
-            else super().to_config(exclude=['_type', '_subtype', '_dict'])
+            else super().to_config(exclude=['_type', '_subtype', '_dict', '_callargs'])
         return {'META': self.meta, 'CONTENT': content() if callable(content) else content}